+++ src/fetchAutocompleteSuggestions.ts
